Overhaul tests
Big change to the test execution, now they are all enclosed in an cljs.test/async block, a macro in
test_helpers.clj that guarantees the correct results.
This involved a huge refactoring that made require-test and source-test (for now only these two)
completely independent from the target).
It also brought up issues before hidden in some test with require-macros, which are behaving in a
weird way, this is why they are confined in weirdo-test and waiting for debugging.